UNOMI-400 More refactoring on property accessors & added unit test
