+++ plugins/baseplugin/src/main/java/org/apache/unomi/plugins/baseplugin/conditions/accessors/ItemAccessor.java
@@ -16,18 +16,29 @@
  */
 package org.apache.unomi.plugins.baseplugin.conditions.accessors;
 
-import org.apache.unomi.api.MetadataItem;
+import org.apache.unomi.api.Item;
+import org.apache.unomi.plugins.baseplugin.conditions.HardcodedPropertyAccessorRegistry;
 
-public class MetadataItemHardcodedPropertyAccessor extends HardcodedPropertyAccessor<MetadataItem> {
-    public MetadataItemHardcodedPropertyAccessor(HardcodedPropertyAccessorRegistry registry) {
+public class ItemAccessor extends HardcodedPropertyAccessor<Item> {
+
+    public ItemAccessor(HardcodedPropertyAccessorRegistry registry) {
         super(registry);
     }
 
     @Override
-    Object getProperty(MetadataItem object, String propertyName, String leftoverExpression) {
-        if ("metadata".equals(propertyName)) {
-            registry.getProperty(object.getMetadata(), leftoverExpression);
+    public Object getProperty(Item object, String propertyName, String leftoverExpression) {
+        if ("itemId".equals(propertyName)) {
+            return object.getItemId();
         }
-        return null;
+        if ("itemType".equals(propertyName)) {
+            return object.getItemType();
+        }
+        if ("scope".equals(propertyName)) {
+            return object.getScope();
+        }
+        if ("version".equals(propertyName)) {
+            return object.getVersion();
+        }
+        return PROPERTY_NOT_FOUND_MARKER;
     }
 }
