+++ plugins/baseplugin/src/main/java/org/apache/unomi/plugins/baseplugin/conditions/accessors/SessionAccessor.java
@@ -16,39 +16,37 @@
  */
 package org.apache.unomi.plugins.baseplugin.conditions.accessors;
 
-import org.apache.unomi.api.Event;
+import org.apache.unomi.api.Session;
+import org.apache.unomi.plugins.baseplugin.conditions.HardcodedPropertyAccessorRegistry;
 
-public class EventHardcodedPropertyAccessor extends HardcodedPropertyAccessor<Event> {
+public class SessionAccessor extends HardcodedPropertyAccessor<Session> {
 
-    public EventHardcodedPropertyAccessor(HardcodedPropertyAccessorRegistry registry) {
+    public SessionAccessor(HardcodedPropertyAccessorRegistry registry) {
         super(registry);
     }
 
     @Override
-    Object getProperty(Event object, String propertyName, String leftoverExpression) {
+    public Object getProperty(Session object, String propertyName, String leftoverExpression) {
+        if ("duration".equals(propertyName)) {
+            return object.getDuration();
+        }
+        if ("size".equals(propertyName)) {
+            return object.getSize();
+        }
+        if ("lastEventDate".equals(propertyName)) {
+            return object.getLastEventDate();
+        }
         if ("properties".equals(propertyName)) {
             return registry.getProperty(object.getProperties(), leftoverExpression);
         }
-        if ("eventType".equals(propertyName)) {
-            return object.getEventType();
+        if ("systemProperties".equals(propertyName)) {
+            return registry.getProperty(object.getSystemProperties(), leftoverExpression);
         }
         if ("profile".equals(propertyName)) {
             return registry.getProperty(object.getProfile(), leftoverExpression);
         }
         if ("profileId".equals(propertyName)) {
             return object.getProfileId();
-        }
-        if ("session".equals(propertyName)) {
-            return registry.getProperty(object.getSession(), leftoverExpression);
-        }
-        if ("sessionId".equals(propertyName)) {
-            return object.getSessionId();
-        }
-        if ("source".equals(propertyName)) {
-            return registry.getProperty(object.getSource(), leftoverExpression);
-        }
-        if ("target".equals(propertyName)) {
-            return registry.getProperty(object.getTarget(), leftoverExpression);
         }
         return PROPERTY_NOT_FOUND_MARKER;
     }
