+++ plugins/baseplugin/src/main/java/org/apache/unomi/plugins/baseplugin/conditions/accessors/TimestampedItemAccessor.java
@@ -16,28 +16,20 @@
  */
 package org.apache.unomi.plugins.baseplugin.conditions.accessors;
 
-import org.apache.unomi.api.Consent;
+import org.apache.unomi.api.TimestampedItem;
+import org.apache.unomi.plugins.baseplugin.conditions.HardcodedPropertyAccessorRegistry;
 
-public class ConsentHardcodedPropertyAccessor extends HardcodedPropertyAccessor<Consent> {
+public class TimestampedItemAccessor extends HardcodedPropertyAccessor<TimestampedItem> {
 
-    public ConsentHardcodedPropertyAccessor(HardcodedPropertyAccessorRegistry registry) {
+    public TimestampedItemAccessor(HardcodedPropertyAccessorRegistry registry) {
         super(registry);
     }
 
     @Override
-    Object getProperty(Consent object, String propertyName, String leftoverExpression) {
-        if ("typeIdentifier".equals(propertyName)) {
-            return object.getTypeIdentifier();
-        } else if ("scope".equals(propertyName)) {
-            return object.getScope();
-        } else if ("status".equals(propertyName)) {
-            return object.getStatus();
-        } else if ("statusDate".equals(propertyName)) {
-            return object.getStatusDate();
-        } else if ("revokeDate".equals(propertyName)) {
-            return object.getRevokeDate();
-        } else {
-            return PROPERTY_NOT_FOUND_MARKER;
+    public Object getProperty(TimestampedItem object, String propertyName, String leftoverExpression) {
+        if ("timeStamp".equals(propertyName)) {
+            return object.getTimeStamp();
         }
+        return PROPERTY_NOT_FOUND_MARKER;
     }
 }
